+++ src/Aschott.Anchor.Application/Cqrs/ICommandHandler.cs
@@ -1,0 +1,7 @@
+using FluentResults;
+using Mediator;
+
+namespace Aschott.Anchor.Application.Cqrs;
+
+public interface ICommandHandler<in TCommand, TResponse> : IRequestHandler<TCommand, Result<TResponse>>
+    where TCommand : ICommand<TResponse>;