+++ src/Aschott.Anchor.Application/Cqrs/IQuery.cs
@@ -1,0 +1,6 @@
+using FluentResults;
+using Mediator;
+
+namespace Aschott.Anchor.Application.Cqrs;
+
+public interface IQuery<TResponse> : IRequest<Result<TResponse>>;